infra/scripts/notion-sync/requirements.txt
notion-client>=2.2.1
python-dotenv>=1.0.0
PyYAML>=6.0.1
